requirements.txt
aiohttp
discord.py
pyyaml
